Math optimizations

+++ Tanki/lighting.h
@@ -12,8 +12,8 @@
 
 
 typedef struct {
-    int16_t x, y;
-    uint16_t distance;
+    float x, y;
+    float distance;
     uint8_t blocked;
 } VisibilityPoint;
